Revert "Revert "refactor(config): reorganize unified-config-types into schemas directory""
This reverts commit 06bce198eb8b9c6508e7b5fc3339cbda54faf608.
